+++ server/client/src/pages/admin/instance-detail-page.tsx
@@ -1,0 +1,5 @@
+const InstanceDetailPage: React.FC = () => {
+  return <></>
+}
+
+export default InstanceDetailPage;
